tipagem e teste com mypy

+++ calculadora-exemplo/principal.py
@@ -22,8 +22,8 @@ while opcao != "5":
         opcao_valida = False
 
     if opcao_valida == True and opcao != "5":
-        numero1 = int(input("Digite o primeiro número: "))
-        numero2 = int(input("Digite o segundo número: "))
+        numero1 = float(input("Digite o primeiro número: "))
+        numero2 = float(input("Digite o segundo número: "))
 
         resultado = 0
         if opcao == '1':
